refactor: move common diagnostic types to a separate module for better organization
